requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.1
aiosignal==1.3.1
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
decouple==0.0.7
dj-database-url==2.3.0
Django==5.1.3
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environs==11.1.0
executing==2.1.0
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
ipython==8.29.0
jedi==0.19.1
marshmallow==3.23.1
matplotlib-inline==0.1.7
multidict==6.1.0
packaging==24.2
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.18.0
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
twilio==9.3.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
wcwidth==0.2.13
whitenoise==6.8.2
yarl==1.17.1
